split out relocator
